Started People Pathfinding
This is before attempting to fix an error in the foreach loop of
Person.cs if the intended fix is ineffective this is the version I will
revert to.

+++ TrafficSim/ZoneMap.cs
@@ -7,7 +7,7 @@ namespace TrafficSim
     {
         public Zone [,] Map;
 
-        public ZoneMap(byte widthByte, byte heightByte)
+        public ZoneMap(int widthByte, int heightByte)
         {
             Map = new Zone[widthByte, heightByte];
         }
